refactor : 경매물품 목 데이터 파일 분리
