+++ src/utils/mocks/api/data/auctionPost/auctionBidRecord.ts
@@ -1,0 +1,3 @@
+import { CommentType } from "../../types";
+
+export const commentData: CommentType[] = [];
